+++ packages/xrpl/HISTORY.md
@@ -10,6 +10,9 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 ### Changed
 * Deprecated `setTransactionFlagsToNumber`. Start using convertTxFlagsToNumber instead
 
+### Added
+* Implementation of XLS-80d PermissionedDomain feature.
+
 ## 4.1.0 (2024-12-23)
 
 ### Added